Move logic for points conversion into viewModel
