+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/chart/ChartModule.kt
@@ -1,13 +1,9 @@
 package io.horizontalsystems.bankwallet.modules.chart
 
-import io.horizontalsystems.bankwallet.core.App
-import io.horizontalsystems.bankwallet.modules.metricchart.MetricChartFactory
-
 object ChartModule {
 
     fun createViewModel(chartService: AbstractChartService): ChartViewModel {
-        val factory = MetricChartFactory(App.numberFormatter)
-        return ChartViewModel(chartService, factory)
+        return ChartViewModel(chartService)
     }
 
 }
