+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/coin/CoinViewFactory.kt
@@ -2,7 +2,6 @@ package io.horizontalsystems.bankwallet.modules.coin
 
 import androidx.annotation.DrawableRes
 import io.horizontalsystems.bankwallet.R
-import io.horizontalsystems.bankwallet.core.App
 import io.horizontalsystems.bankwallet.core.IAppNumberFormatter
 import io.horizontalsystems.bankwallet.core.providers.Translator
 import io.horizontalsystems.bankwallet.entities.CurrencyValue
@@ -10,15 +9,12 @@ import io.horizontalsystems.bankwallet.entities.order
 import io.horizontalsystems.bankwallet.modules.coin.overview.CoinOverviewItem
 import io.horizontalsystems.bankwallet.modules.coin.overview.CoinOverviewViewItem
 import io.horizontalsystems.chartview.ChartData
-import io.horizontalsystems.chartview.ChartDataFactory
 import io.horizontalsystems.chartview.ChartView
-import io.horizontalsystems.chartview.models.ChartPoint
 import io.horizontalsystems.chartview.models.MacdInfo
 import io.horizontalsystems.core.entities.Currency
 import io.horizontalsystems.core.helpers.DateHelper
 import io.horizontalsystems.marketkit.models.*
 import io.horizontalsystems.views.ListPosition
-import java.lang.Long.max
 import java.math.BigDecimal
 import java.net.URI
 
@@ -131,29 +127,6 @@ class CoinViewFactory(
     private val currency: Currency,
     private val numberFormatter: IAppNumberFormatter
 ) {
-    fun createChartInfoData(
-        type: ChartType,
-        chartInfo: ChartInfo,
-        lastPoint: LastPoint
-    ): ChartInfoData {
-        val chartType = when (type) {
-            ChartType.TODAY -> ChartView.ChartType.TODAY
-            ChartType.DAILY -> ChartView.ChartType.DAILY
-            ChartType.WEEKLY -> ChartView.ChartType.WEEKLY
-            ChartType.WEEKLY2 -> ChartView.ChartType.WEEKLY2
-            ChartType.MONTHLY -> ChartView.ChartType.MONTHLY
-            ChartType.MONTHLY_BY_DAY -> ChartView.ChartType.MONTHLY_BY_DAY
-            ChartType.MONTHLY3 -> ChartView.ChartType.MONTHLY3
-            ChartType.MONTHLY6 -> ChartView.ChartType.MONTHLY6
-            ChartType.MONTHLY12 -> ChartView.ChartType.MONTHLY12
-            ChartType.MONTHLY24 -> ChartView.ChartType.MONTHLY24
-        }
-        val chartData = createChartData(chartInfo, lastPoint, chartType)
-        val maxValue = numberFormatter.formatFiat(chartData.valueRange.upper, currency.symbol, 0, 2)
-        val minValue = numberFormatter.formatFiat(chartData.valueRange.lower, currency.symbol, 0, 2)
-
-        return ChartInfoData(chartData, chartType, maxValue, minValue)
-    }
 
     fun getRoi(performance: Map<String, Map<TimePeriod, BigDecimal>>): List<RoiViewItem> {
         val rows = mutableListOf<RoiViewItem>()
@@ -323,11 +296,6 @@ class CoinViewFactory(
         return links
     }
 
-    fun getFormattedLatestRate(currencyValue: CurrencyValue): String {
-        val significantDecimal = App.numberFormatter.getSignificantDecimalFiat(currencyValue.value)
-        return numberFormatter.formatFiat(currencyValue.value, currencyValue.currency.symbol, 2, significantDecimal)
-    }
-
     private fun getIcon(linkType: LinkType): Int {
         return when (linkType) {
             LinkType.Guide -> R.drawable.ic_academy_20
@@ -359,44 +327,6 @@ class CoinViewFactory(
         }
     }
 
-    private fun createChartData(
-        chartInfo: ChartInfo,
-        lastPoint: LastPoint,
-        chartType: ChartView.ChartType
-    ): ChartData {
-        val points = chartInfo.points.map {
-            ChartPoint(
-                it.value.toFloat(),
-                it.volume?.toFloat(),
-                it.timestamp
-            )
-        }.toMutableList()
-        val chartInfoLastPoint = chartInfo.points.lastOrNull()
-        var endTimestamp = chartInfo.endTimestamp
-
-        if (chartInfoLastPoint?.timestamp != null && lastPoint.timestamp > chartInfoLastPoint.timestamp) {
-            endTimestamp = max(lastPoint.timestamp, endTimestamp)
-            points.add(ChartPoint(lastPoint.rate.toFloat(), null, lastPoint.timestamp))
-
-            if (chartType == ChartView.ChartType.DAILY) {
-                val startTimestamp = lastPoint.timestamp - 24 * 60 * 60
-                val startValue =
-                    lastPoint.rate.multiply(100.toBigDecimal()) / lastPoint.rateDiff24h.add(100.toBigDecimal())
-                val startPoint = ChartPoint(startValue.toFloat(), null, startTimestamp)
-
-                points.removeIf { it.timestamp <= startTimestamp }
-                points.add(0, startPoint)
-            }
-        }
-
-        return ChartDataFactory.build(
-            points,
-            chartInfo.startTimestamp,
-            endTimestamp,
-            chartInfo.isExpired
-        )
-    }
-
     private fun formatFiatShortened(value: BigDecimal, symbol: String): String {
         val shortCapValue = numberFormatter.shortenValue(value)
         return numberFormatter.formatFiat(
@@ -408,14 +338,3 @@ class CoinViewFactory(
     }
 
 }
-
-data class MarketDataViewItem(
-    val marketCap: String?,
-    val marketCapRank: String?,
-    val volume24h: String?,
-    val tvl: String?,
-    val genesisDate: String?,
-    val circulatingSupply: String?,
-    val totalSupply: String?,
-    val dilutedMarketCap: String?,
-)
